app/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.27.0
google-cloud==0.34.0
google-cloud-compute==1.15.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
multidict==6.0.4
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.26.0
rsa==4.9
urllib3==1.26.18
yarl==1.9.4